requirements.txt
google-genai>=1.49.0
python-dotenv>=1.0.0